requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-tavily
langgraph
youtube-transcript-api
python-dotenv
